chore(ci): Add linter step
