+++ .github/workflows/test.yml
@@ -19,6 +19,8 @@ jobs:
         uses: actions/setup-go@v5
         with:
           go-version: "^1.23"
+      - name: Run golangci-lint
+        uses: golangci/golangci-lint-action@v8.0.0
       - name: Check formatting
         run: |
           unformatted=$(gofmt -l .)
